## User Entity 수정, 패키지 변경
### Description
- User Entity 수정
- domain 패키지 내로 이동, User 클래스에 의존하는 다른 Entity import 문 수정
