+++ src/main/java/com/blog/tanylog/postLike/PostLike.java
@@ -1,8 +1,7 @@
 package com.blog.tanylog.postLike;
 
 import com.blog.tanylog.post.Post;
-import com.blog.tanylog.user.User;
-import javax.persistence.Column;
+import com.blog.tanylog.user.domain.User;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
